♻️ refactor[prisma]: update import paths for prisma client in service files
